Add multer for file upload and implement getMySongs endpoint

+++ controllers/song.js
@@ -27,9 +27,9 @@ const deleteSong = async (req, res) => {
     if (!song) {
         return res.status(404).json({ message: 'Song not found' })
     }
-    // if (song.createdBy.toString() !== req.user._id.toString()) {
-    //     return res.status(403).json({ message: 'Not authorized' })
-    // }
+    if (song.createdBy.toString() !== req.user._id.toString()) {
+        return res.status(403).json({ message: 'Not authorized' })
+    }
     fs.unlinkSync(song.file)
     await Song.deleteOne({ _id: id })
 
@@ -38,6 +38,13 @@ const deleteSong = async (req, res) => {
 
 const getAllSongs = async (req, res) => {
     const songs = await Song.find()
+    res.status(200).json({ songs })
+}
+
+const getMySongs = async (req, res) => {
+    const createdBy = req.user._id
+    // console.log('songs', createdBy)
+    const songs = await Song.find({ createdBy })
     res.status(200).json({ songs })
 }
 const getSong = async (req, res) => {
@@ -50,6 +57,7 @@ const getSong = async (req, res) => {
 }
 const updateSong = async (req, res) => {
     const { id } = req.params
+    // console.log('req.body', req.body)
     const { title, artist, description } = req.body
     const song = await Song.findById(id)
     if (!song) {
@@ -62,7 +70,7 @@ const updateSong = async (req, res) => {
     song.artist = artist || song.artist
     song.description = description || song.description
     await song.save()
-    res.status(200).json({ message: 'updateSong' })
+    res.status(200).json({ song })
 }
 
 const getSongFile = async (req, res) => {
@@ -81,4 +89,5 @@ module.exports = {
     getSong,
     updateSong,
     getSongFile,
+    getMySongs,
 }
